sliced volume rendering
render grass in the middle distance as a set of axis aligned vertical slices.
On the first frame rander orthographic slices of the full grass patch mesh to
the texture to be used for the slices. On the left half of the texture write
the lit output, on the right half write the density attribute. Use the density
attribute to perform distance based discarding.

There's a slightly incorrect behaviour, both x axis aligned and z axis aligned
slices use the same texture. I'm not sure this is actually visible in the
output.